+++ CHANGELOG.md
@@ -25,6 +25,8 @@ Line wrap the file at 100 chars.                                              Th
 ## [Unreleased]
 ### Added
 - Add CLI command to set the location constraint via `mullvad relay set relay HOSTNAME`.
+- Add a provider relay constraint, which restricts relay selection to a given hosting provider.
+- Include hosting providers in the CLI for `mullvad relay list` and `mullvad bridge list`.
 
 ### Changed
 - Use gRPC for communication between frontends and the backend instead of JSON-RPC.